back end/requirements.txt
astroid==2.5.6
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
peewee==3.14.8
pylint==2.8.3
python-dotenv==0.19.2
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
wrapt==1.12.1
